EcoSensors/requirements.txt
asgiref==3.5.2
django==4.0.6
sqlparse==0.4.2
djangorestframework==3.13.1
djangorestframework-gis==1
django-filter==22.1.0
